# openapi.yml
openapi: "3.0.0"
info:
  version: v0
  title: youyuan core Service
  license:
    name: MIT
servers:
  - url: https://api.youyuan.36node.com{basePath}
    description: The production API server
    variables:
      basePath:
        default: /core/v0
paths:
...
  /drafts:
    post:
      summary: Create a draft
      operationId: createDraft
      tags:
        - draft
      requestBody:
        description: draft to add
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftCreateDoc"
      responses:
        "201":
          description: The draft created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      summary: List all drafts
      operationId: listDafts
      tags:
        - draft
      parameters:
        - name: _limit
          in: query
          description: How many items to return at one time (max 100)
          required: false
          schema:
            type: integer
            format: int32
            default: 10
            maximum: 1000
        - name: _offset
          in: query
          description: How many items to escape
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: _sort
          in: query
          description: Fields to sort
          required: false
          schema:
            type: string
        - name: _select
          in: query
          description: Fields to select
          required: false
          schema:
            type: array
            items:
              type: string
        - name: type
          in: query
          description: draft type
          required: true
          schema:
            type: string
            enum: ["VEHICLE", "REPAIR", "APPEARANCE", "SOFTWARE"]
        - name: resource
          in: query
          description: 可按照维保关联的资源查询
          schema:
            type: string
        - name: ns_like
          in: query
          description: ns like
          schema:
            type: string
        - name: createBy
          in: query
          description: createBy
          schema:
            type: string
      responses:
        "200":
          description: A paged array of drafts
          headers:
            X-Total-Count:
              description: Total count of all warnings match the query
              required: true
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Draft"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /drafts/{draftId}:
    get:
      summary: Find draft by id
      operationId: getDraft
      tags:
        - draft
      parameters:
        - name: draftId
          in: path
          required: true
          description: The id of the draft to retrieve
          schema:
            type: string
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      summary: Update draft
      operationId: updateDraft
      tags:
        - draft
      parameters:
        - name: draftId
          in: path
          required: true
          description: The id of draft
          schema:
            type: string
      requestBody:
        description: draft to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftDoc"
      responses:
        "200":
          description: The draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      description: deletes a single draft based on the ID supplied
      operationId: deleteDraft
      tags:
        - draft
      parameters:
        - name: draftId
          in: path
          description: ID of draft to delete
          required: true
          schema:
            type: string
      responses:
        "204":
          description: draft deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
security:
  - bearerAuth: []
components:
  responses:
...
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
    InternalError:
      description: unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
  schemas:
...
    DraftDoc:
      description: 草稿箱信息
      type: object
      properties:
        ns:
          type: string
          description: 草稿所属ns
        resource:
          type: string
          description: 关联资源 id
        type:
          type: string
          enum: ["VEHICLE", "REPAIR", "MAINTAIN"]
          description: 草稿类型
        content:
          type: string
          description: 草稿箱内容
        createBy:
          type: string
          description: 创建用户id
    DraftCreateDoc:
      allOf:
        - $ref: "#/components/schemas/DraftDoc"
        - type: object
          required:
            - type
            - ns
            - createBy
          properties:
            type:
              type: string
              enum: ["VEHICLE", "REPAIR", "MAINTAIN"]
            ns:
              type: string
            createBy:
              type: string
    Draft:
      allOf:
        - $ref: "#/components/schemas/DraftDoc"
        - $ref: "#/components/schemas/MongoDefault"
...
    MongoDefault:
      type: object
      required:
        - id
      properties:
        id:
          type: string
        updateAt:
          type: string
          format: date-time
        updateBy:
          type: string
        createAt:
          type: string
          format: date-time
        createBy:
          type: string
    Err:
      type: object
      required:
        - name
        - message
      properties:
        code:
          type: string
        type:
          type: string
        message:
          type: boolean
        name:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              keyword:
                type: string
              message:
                type: string
              path:
                type: string
              value:
                type: string
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT